+++ src/systems/phone/esp_ui_phone_app.cpp
@@ -67,15 +67,15 @@ bool ESP_UI_PhoneApp::beginExtra(void)
     _active_data = _init_data;
 
     // Check navigation bar and gesture
-    if ((_active_data.navigation_bar_visual_mode != ESP_UI_PHONE_APP_VISUAL_MODE_HIDE) &&
+    if ((_active_data.navigation_bar_visual_mode != ESP_UI_NAVIGATION_BAR_VISUAL_MODE_HIDE) &&
             _active_data.flags.enable_navigation_gesture) {
         ESP_UI_LOGW("Both navigation bar and gesture are enabled, only bar will be used");
         _active_data.flags.enable_navigation_gesture = false;
     }
-    if ((_active_data.navigation_bar_visual_mode != ESP_UI_PHONE_APP_VISUAL_MODE_HIDE) &&
+    if ((_active_data.navigation_bar_visual_mode != ESP_UI_NAVIGATION_BAR_VISUAL_MODE_HIDE) &&
             (navigation_bar == nullptr)) {
         ESP_UI_LOGE("Navigation bar is enabled but not provided");
-        _active_data.navigation_bar_visual_mode = ESP_UI_PHONE_APP_VISUAL_MODE_HIDE;
+        _active_data.navigation_bar_visual_mode = ESP_UI_NAVIGATION_BAR_VISUAL_MODE_HIDE;
     } else if (_active_data.flags.enable_navigation_gesture && (gesture == nullptr)) {
         ESP_UI_LOGE("Navigation gesture is enabled but not provided");
         _active_data.flags.enable_navigation_gesture = false;