+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # ChangeLog
 
-## v0.1.0 - 2024-08-09
+## v0.1.0 - 2024-08-12
 
 ### Enhancements:
 
 * feat(repo): init project
-* feat(systems): add "Phone" system
+* feat(systems): add 'Phone' system
+* feat(widgets): support 'Status Bar' and 'Navigation Bar' visual mode